requirements.txt
django>=1.8.0
mysqlclient>=1.3.6
salt>=2014.7.5
celery>=3.1
cmdb-sdk
djangorestframework
markdown
django-filter
django-celery>=3.1

